Refactor settings into tabbed layout
