+++ src/app/routes/SettingsPage.tsx
@@ -1,9 +1,9 @@
-import { useEffect, useMemo, useState } from 'react';
+import { useEffect, useMemo, useState, type FormEvent } from 'react';
 import { useSettings } from '../state/SettingsContext';
 import type { ThemePreference, WeekStart, Weekday } from '../db/schema';
 import { fetchPublicHolidays, fetchPublicHolidayRegions, type HolidayRegion } from '../logic/publicHolidays';
-import ImportExportModal from '../components/ImportExportModal';
-import BackupRestoreModal from '../components/BackupRestoreModal';
+import ImportExportPanel from '../components/ImportExportPanel';
+import BackupRestorePanel from '../components/BackupRestorePanel';
 
 const WEEK_START_OPTIONS: Array<{ value: WeekStart; label: string }> = [
   { value: 0, label: 'Sunday' },
@@ -38,6 +38,36 @@ const THEME_OPTIONS: Array<{ value: ThemePreference; label: string; description:
   { value: 'light', label: 'Light', description: 'Always use the light theme.' },
   { value: 'dark', label: 'Dark', description: 'Always use the dark theme.' }
 ];
+
+const SETTINGS_TABS = [
+  {
+    id: 'general',
+    label: 'General',
+    description: 'Base rate, currency, and week alignment.'
+  },
+  {
+    id: 'notifications',
+    label: 'Notifications',
+    description: 'Control reminder lead times.'
+  },
+  {
+    id: 'penalties',
+    label: 'Penalty rules',
+    description: 'Daily windows, weekends, and public holidays.'
+  },
+  {
+    id: 'appearance',
+    label: 'Appearance',
+    description: 'Theme and time display preferences.'
+  },
+  {
+    id: 'data',
+    label: 'Data & backup',
+    description: 'Import/export shifts and manage archives.'
+  }
+] as const;
+
+type SettingsTabId = (typeof SETTINGS_TABS)[number]['id'];
 
 function minutesToTime(minutes: number): string {
   const safeMinutes = Number.isFinite(minutes) ? Math.max(0, Math.min(minutes, 24 * 60)) : 0;
@@ -93,6 +123,7 @@ export function timeToMinutes(value: string): number {
 
 export default function SettingsPage() {
   const { settings, updateSettings, isLoading, error } = useSettings();
+  const [activeTab, setActiveTab] = useState<SettingsTabId>('general');
   const [baseRate, setBaseRate] = useState(() => settings?.baseRate ?? 25);
   const [penaltyRate, setPenaltyRate] = useState(() => settings?.penaltyRate ?? 35);
   const [weekStartsOn, setWeekStartsOn] = useState<WeekStart>(() => settings?.weekStartsOn ?? 1);
@@ -125,8 +156,11 @@ export default function SettingsPage() {
   const [status, setStatus] = useState<string | null>(null);
   const [formError, setFormError] = useState<string | null>(null);
   const [isSaving, setIsSaving] = useState(false);
-  const [isImportExportOpen, setIsImportExportOpen] = useState(false);
-  const [isBackupModalOpen, setIsBackupModalOpen] = useState(false);
+
+  useEffect(() => {
+    setFormError(null);
+    setStatus(null);
+  }, [activeTab]);
 
   useEffect(() => {
     if (settings) {
@@ -220,6 +254,102 @@ export default function SettingsPage() {
     return name ?? publicHolidaySubdivision;
   }, [holidayRegions, publicHolidaySubdivision]);
 
+  const handleSubmit = async (event: FormEvent<HTMLFormElement>) => {
+    event.preventDefault();
+    setFormError(null);
+    setStatus(null);
+    setIsSaving(true);
+
+    const startMinutes = timeToMinutes(penaltyStartTime);
+    const endMinutes = timeToMinutes(penaltyEndTime);
+
+    if (penaltyDailyWindowEnabled && endMinutes <= startMinutes) {
+      setFormError('Penalty end time must be after the start time.');
+      setIsSaving(false);
+      return;
+    }
+
+    const selectedDays = Array.from(new Set(penaltyAllDayWeekdays)).sort((a, b) => a - b) as Weekday[];
+
+    const normalizedNotificationLongLead = Math.max(0, Math.floor(Number(notificationLongLead) || 0));
+    const normalizedNotificationShortLead = Math.max(0, Math.floor(Number(notificationShortLead) || 0));
+    const normalizedNotificationRepeat = Math.max(5, Math.floor(Number(notificationRepeat) || 0));
+
+    let publicHolidayDates = includePublicHolidays ? settings?.publicHolidayDates ?? [] : [];
+    const normalizedSubdivision = publicHolidaySubdivision.trim().toUpperCase();
+    const countryChanged = settings?.publicHolidayCountry !== publicHolidayCountry.toUpperCase();
+    const subdivisionChanged = (settings?.publicHolidaySubdivision ?? '') !== normalizedSubdivision;
+    const includeChanged = settings?.includePublicHolidays !== includePublicHolidays;
+
+    if (
+      includePublicHolidays &&
+      (publicHolidayDates.length === 0 || countryChanged || includeChanged || subdivisionChanged)
+    ) {
+      try {
+        const currentYear = new Date().getFullYear();
+        publicHolidayDates = await fetchPublicHolidays(
+          publicHolidayCountry,
+          [currentYear - 1, currentYear, currentYear + 1],
+          normalizedSubdivision
+        );
+      } catch (holidayError) {
+        setFormError(
+          holidayError instanceof Error
+            ? holidayError.message
+            : 'Failed to load public holiday data. Please try again later.'
+        );
+        setIsSaving(false);
+        return;
+      }
+    }
+
+    try {
+      await updateSettings({
+        baseRate: Number(baseRate),
+        penaltyRate: Number(penaltyRate),
+        weekStartsOn,
+        currency,
+        theme,
+        use24HourTime,
+        notificationLongLeadMinutes: normalizedNotificationLongLead,
+        notificationShortLeadMinutes: normalizedNotificationShortLead,
+        notificationRepeatMinutes: normalizedNotificationRepeat,
+        penaltyDailyWindowEnabled,
+        penaltyDailyStartMinute: startMinutes,
+        penaltyDailyEndMinute: endMinutes,
+        penaltyAllDayWeekdays: selectedDays,
+        includePublicHolidays,
+        publicHolidayCountry: publicHolidayCountry.toUpperCase(),
+        publicHolidaySubdivision: normalizedSubdivision,
+        publicHolidayDates
+      });
+      setStatus('Settings saved');
+      setTimeout(() => setStatus(null), 2500);
+    } catch (settingsError) {
+      setFormError(
+        settingsError instanceof Error
+          ? settingsError.message
+          : 'Failed to save settings. Please try again.'
+      );
+    } finally {
+      setIsSaving(false);
+    }
+  };
+
+  const renderFormActions = () => (
+    <div className="flex flex-col gap-2">
+      <button
+        type="submit"
+        className="w-full rounded-full bg-primary px-5 py-2 text-sm font-semibold text-primary-foreground shadow transition hover:bg-midnight-900 disabled:cursor-not-allowed disabled:opacity-60 sm:w-auto"
+        disabled={isSaving}
+      >
+        {isSaving ? 'Saving…' : 'Save settings'}
+      </button>
+      {formError && <p className="text-xs text-rose-500">{formError}</p>}
+      {status && <p className="text-xs text-emerald-500">{status}</p>}
+    </div>
+  );
+
   if (isLoading) {
     return <p className="text-sm text-neutral-500">Chrona is loading your preferences…</p>;
   }
@@ -229,408 +359,360 @@ export default function SettingsPage() {
   }
 
   return (
-    <section className="max-w-xl rounded-2xl border border-neutral-200 bg-white p-6 shadow-sm dark:border-midnight-800 dark:bg-midnight-900">
-      <h2 className="text-lg font-semibold text-neutral-900 dark:text-neutral-50">Chrona preferences</h2>
-      <p className="mb-4 text-sm text-neutral-500 dark:text-neutral-300">
-        Tune pay rates, notifications, and penalty windows so Chrona mirrors the way you work.
-      </p>
-      <div className="mb-6 rounded-xl border border-neutral-200 bg-neutral-100/60 p-4 dark:border-midnight-800 dark:bg-midnight-900/60">
-        <h3 className="text-sm font-semibold text-neutral-900 dark:text-neutral-100">Data management</h3>
-        <p className="mt-1 text-xs text-neutral-500 dark:text-neutral-300">
-          Backup or export your data, or restore and import from a saved file.
-        </p>
-        <div className="mt-3 flex flex-wrap gap-2">
-          <button
-            type="button"
-            onClick={() => setIsImportExportOpen(true)}
-            className="inline-flex items-center rounded-full border border-neutral-300 px-3 py-1.5 text-sm font-medium text-neutral-700 transition-colors hover:bg-neutral-100 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-primary focus-visible:ring-offset-2 dark:border-midnight-700 dark:text-neutral-100 dark:hover:bg-midnight-800 dark:focus-visible:ring-offset-midnight-900"
-          >
-            Import / Export
-          </button>
-          <button
-            type="button"
-            onClick={() => setIsBackupModalOpen(true)}
-            className="inline-flex items-center rounded-full border border-neutral-300 px-3 py-1.5 text-sm font-medium text-neutral-700 transition-colors hover:bg-neutral-100 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-primary focus-visible:ring-offset-2 dark:border-midnight-700 dark:text-neutral-100 dark:hover:bg-midnight-800 dark:focus-visible:ring-offset-midnight-900"
-          >
-            Backup / Restore
-          </button>
-        </div>
-      </div>
-      <form
-        className="flex flex-col gap-5"
-        onSubmit={async (event) => {
-          event.preventDefault();
-          setFormError(null);
-          setStatus(null);
-          setIsSaving(true);
-
-          const startMinutes = timeToMinutes(penaltyStartTime);
-          const endMinutes = timeToMinutes(penaltyEndTime);
-
-          if (penaltyDailyWindowEnabled && endMinutes <= startMinutes) {
-            setFormError('Penalty end time must be after the start time.');
-            setIsSaving(false);
-            return;
-          }
-
-          const selectedDays = Array.from(new Set(penaltyAllDayWeekdays)).sort((a, b) => a - b) as Weekday[];
-
-          const normalizedNotificationLongLead = Math.max(0, Math.floor(Number(notificationLongLead) || 0));
-          const normalizedNotificationShortLead = Math.max(0, Math.floor(Number(notificationShortLead) || 0));
-          const normalizedNotificationRepeat = Math.max(5, Math.floor(Number(notificationRepeat) || 0));
-
-          let publicHolidayDates = includePublicHolidays ? settings?.publicHolidayDates ?? [] : [];
-          const normalizedSubdivision = publicHolidaySubdivision.trim().toUpperCase();
-          const countryChanged = settings?.publicHolidayCountry !== publicHolidayCountry.toUpperCase();
-          const subdivisionChanged = (settings?.publicHolidaySubdivision ?? '') !== normalizedSubdivision;
-          const includeChanged = settings?.includePublicHolidays !== includePublicHolidays;
-
-          if (
-            includePublicHolidays &&
-            (publicHolidayDates.length === 0 || countryChanged || includeChanged || subdivisionChanged)
-          ) {
-            try {
-              const currentYear = new Date().getFullYear();
-              publicHolidayDates = await fetchPublicHolidays(
-                publicHolidayCountry,
-                [currentYear - 1, currentYear, currentYear + 1],
-                normalizedSubdivision
-              );
-            } catch (holidayError) {
-              setFormError(
-                holidayError instanceof Error
-                  ? holidayError.message
-                  : 'Failed to load public holiday data. Please try again later.'
-              );
-              setIsSaving(false);
-              return;
-            }
-          }
-
-          try {
-            await updateSettings({
-              baseRate: Number(baseRate),
-              penaltyRate: Number(penaltyRate),
-              weekStartsOn,
-              currency,
-              theme,
-              use24HourTime,
-              notificationLongLeadMinutes: normalizedNotificationLongLead,
-              notificationShortLeadMinutes: normalizedNotificationShortLead,
-              notificationRepeatMinutes: normalizedNotificationRepeat,
-              penaltyDailyWindowEnabled,
-              penaltyDailyStartMinute: startMinutes,
-              penaltyDailyEndMinute: endMinutes,
-              penaltyAllDayWeekdays: selectedDays,
-              includePublicHolidays,
-              publicHolidayCountry: publicHolidayCountry.toUpperCase(),
-              publicHolidaySubdivision: normalizedSubdivision,
-              publicHolidayDates
-            });
-            setStatus('Settings saved');
-            setTimeout(() => setStatus(null), 2500);
-          } catch (settingsError) {
-            setFormError(
-              settingsError instanceof Error
-                ? settingsError.message
-                : 'Failed to save settings. Please try again.'
-            );
-          } finally {
-            setIsSaving(false);
-          }
-        }}
-      >
-        <div className="grid gap-2">
-          <label className="text-xs font-semibold uppercase text-neutral-500">Base rate (per hour)</label>
-          <input
-            type="number"
-            min="0"
-            step="0.01"
-            value={baseRate}
-            onChange={(event) => setBaseRate(Number(event.target.value))}
-            className="rounded-lg border border-neutral-200 bg-white px-3 py-2 text-sm shadow-sm focus:border-primary focus:outline-none focus:ring-2 focus:ring-primary/40 dark:border-midnight-700 dark:bg-midnight-900"
-          />
-        </div>
-        <div className="grid gap-2">
-          <label className="text-xs font-semibold uppercase text-neutral-500">Penalty rate (per hour)</label>
-          <input
-            type="number"
-            min="0"
-            step="0.01"
-            value={penaltyRate}
-            onChange={(event) => setPenaltyRate(Number(event.target.value))}
-            className="rounded-lg border border-neutral-200 bg-white px-3 py-2 text-sm shadow-sm focus:border-primary focus:outline-none focus:ring-2 focus:ring-primary/40 dark:border-midnight-700 dark:bg-midnight-900"
-          />
-        </div>
-        <div className="grid gap-2">
-          <label
-            className="text-xs font-semibold uppercase text-neutral-500"
-            htmlFor="week-starts-on"
-          >
-            Pay week starts on
-          </label>
-          <p id="week-starts-on-help" className="text-xs text-neutral-500 dark:text-neutral-300">
-            Only affects summary and payslip alignment.
+    <section className="mx-auto flex max-w-5xl flex-col gap-6 rounded-2xl border border-neutral-200 bg-white p-6 shadow-sm dark:border-midnight-800 dark:bg-midnight-900">
+      <div className="flex flex-col gap-6 lg:flex-row">
+        <aside className="lg:w-72">
+          <h2 className="text-lg font-semibold text-neutral-900 dark:text-neutral-50">Chrona preferences</h2>
+          <p className="mt-1 text-sm text-neutral-500 dark:text-neutral-300">
+            Tune pay rates, notifications, and penalty windows so Chrona mirrors the way you work.
           </p>
-          <select
-            id="week-starts-on"
-            aria-describedby="week-starts-on-help"
-            value={weekStartsOn}
-            onChange={(event) => setWeekStartsOn(Number(event.target.value) as WeekStart)}
-            className="rounded-lg border border-neutral-200 bg-white px-3 py-2 text-sm shadow-sm focus:border-primary focus:outline-none focus:ring-2 focus:ring-primary/40 dark:border-midnight-700 dark:bg-midnight-900"
-          >
-            {WEEK_START_OPTIONS.map((option) => (
-              <option key={option.value} value={option.value}>
-                {option.label}
-              </option>
-            ))}
-          </select>
-        </div>
-        <div className="grid gap-2">
-          <label className="text-xs font-semibold uppercase text-neutral-500">Currency</label>
-          <input
-            type="text"
-            value={currency}
-            onChange={(event) => setCurrency(event.target.value.toUpperCase())}
-            className="rounded-lg border border-neutral-200 bg-white px-3 py-2 text-sm uppercase shadow-sm focus:border-primary focus:outline-none focus:ring-2 focus:ring-primary/40 dark:border-midnight-700 dark:bg-midnight-900"
-            maxLength={3}
-          />
-        </div>
-        <div className="grid gap-2">
-          <span className="text-xs font-semibold uppercase text-neutral-500">Time format</span>
-          <label className="flex items-center gap-2 text-sm text-neutral-600 dark:text-neutral-200">
-            <input
-              type="checkbox"
-              checked={use24HourTime}
-              onChange={(event) => setUse24HourTime(event.target.checked)}
-              className="h-4 w-4 rounded border-neutral-300 text-primary focus:ring-primary"
-            />
-            Use 24-hour clock
-          </label>
-          <p className="text-xs text-neutral-500 dark:text-neutral-300">
-            Applies to shift start and end times shown in the app.
-          </p>
-        </div>
-        <fieldset className="grid gap-3">
-          <legend className="text-xs font-semibold uppercase text-neutral-500">Shift reminders</legend>
-          <p className="text-xs text-neutral-500 dark:text-neutral-300">
-            Configure how far in advance the app reminds you about upcoming shifts.
-          </p>
-          <div className="grid gap-3 sm:grid-cols-3">
-            <label className="grid gap-1 text-sm text-neutral-600 dark:text-neutral-200">
-              <span className="text-xs font-semibold uppercase text-neutral-500">Long-range (minutes)</span>
-              <input
-                type="number"
-                min={0}
-                max={7 * 24 * 60}
-                value={notificationLongLead}
-                onChange={(event) => setNotificationLongLead(Number(event.target.value))}
-                className="rounded-lg border border-neutral-200 bg-white px-3 py-2 text-sm shadow-sm focus:border-primary focus:outline-none focus:ring-2 focus:ring-primary/40 dark:border-midnight-700 dark:bg-midnight-900"
-              />
-              <span className="text-xs text-neutral-500 dark:text-neutral-300">
-                Send a one-off reminder this many minutes before the shift.
-              </span>
-            </label>
-            <label className="grid gap-1 text-sm text-neutral-600 dark:text-neutral-200">
-              <span className="text-xs font-semibold uppercase text-neutral-500">Short-range (minutes)</span>
-              <input
-                type="number"
-                min={0}
-                max={24 * 60}
-                value={notificationShortLead}
-                onChange={(event) => setNotificationShortLead(Number(event.target.value))}
-                className="rounded-lg border border-neutral-200 bg-white px-3 py-2 text-sm shadow-sm focus:border-primary focus:outline-none focus:ring-2 focus:ring-primary/40 dark:border-midnight-700 dark:bg-midnight-900"
-              />
-              <span className="text-xs text-neutral-500 dark:text-neutral-300">
-                Start persistent reminders once the shift is this close.
-              </span>
-            </label>
-            <label className="grid gap-1 text-sm text-neutral-600 dark:text-neutral-200">
-              <span className="text-xs font-semibold uppercase text-neutral-500">Repeat every (minutes)</span>
-              <input
-                type="number"
-                min={5}
-                max={24 * 60}
-                value={notificationRepeat}
-                onChange={(event) => setNotificationRepeat(Number(event.target.value))}
-                className="rounded-lg border border-neutral-200 bg-white px-3 py-2 text-sm shadow-sm focus:border-primary focus:outline-none focus:ring-2 focus:ring-primary/40 dark:border-midnight-700 dark:bg-midnight-900"
-              />
-              <span className="text-xs text-neutral-500 dark:text-neutral-300">
-                How often short-range reminders repeat until the shift starts.
-              </span>
-            </label>
+          <div className="mt-4 rounded-xl border border-neutral-200 bg-neutral-100/70 p-2 dark:border-midnight-800 dark:bg-midnight-900/60">
+            <div className="flex gap-2 overflow-x-auto pb-2 lg:flex-col lg:overflow-visible lg:pb-0">
+              {SETTINGS_TABS.map((tab) => {
+                const isActive = activeTab === tab.id;
+                const buttonClasses = [
+                  'min-w-[160px] rounded-lg px-3 py-2 text-left transition focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-primary focus-visible:ring-offset-2 dark:focus-visible:ring-offset-midnight-900 lg:min-w-0',
+                  isActive ? 'bg-white shadow-sm ring-1 ring-primary/60 dark:bg-midnight-950' : 'hover:bg-white/70 dark:hover:bg-midnight-800/80'
+                ].join(' ');
+                const labelClasses = isActive
+                  ? 'block text-sm font-semibold text-neutral-900 dark:text-neutral-50'
+                  : 'block text-sm font-semibold text-neutral-700 dark:text-neutral-200';
+                const descriptionClasses = 'mt-1 block text-xs text-neutral-500 dark:text-neutral-400';
+                return (
+                  <button
+                    key={tab.id}
+                    type="button"
+                    onClick={() => setActiveTab(tab.id)}
+                    className={buttonClasses}
+                    aria-current={isActive ? 'page' : undefined}
+                  >
+                    <span className={labelClasses}>{tab.label}</span>
+                    <span className={descriptionClasses}>{tab.description}</span>
+                  </button>
+                );
+              })}
+            </div>
           </div>
-        </fieldset>
-        <fieldset className="grid gap-3">
-          <legend className="text-xs font-semibold uppercase text-neutral-500">Appearance</legend>
-          <div className="grid gap-2">
-            {THEME_OPTIONS.map((option) => (
-              <label
-                key={option.value}
-                className="flex items-start gap-2 rounded-lg border border-neutral-200 bg-white px-3 py-2 text-sm shadow-sm transition hover:border-primary/60 dark:border-midnight-700 dark:bg-midnight-900"
-              >
-                <input
-                  type="radio"
-                  name="theme"
-                  value={option.value}
-                  checked={theme === option.value}
-                  onChange={() => setTheme(option.value)}
-                  className="mt-1 h-4 w-4 border-neutral-300 text-primary focus:ring-primary"
-                />
-                <span className="flex flex-col">
-                  <span className="font-medium text-neutral-700 dark:text-neutral-100">{option.label}</span>
-                  <span className="text-xs text-neutral-500 dark:text-neutral-300">{option.description}</span>
-                </span>
-              </label>
-            ))}
-          </div>
-        </fieldset>
-        <fieldset className="grid gap-3">
-          <legend className="text-xs font-semibold uppercase text-neutral-500">Penalty hours (daily window)</legend>
-          <label className="flex items-center gap-2 text-sm text-neutral-600 dark:text-neutral-200">
-            <input
-              type="checkbox"
-              checked={penaltyDailyWindowEnabled}
-              onChange={(event) => setPenaltyDailyWindowEnabled(event.target.checked)}
-              className="h-4 w-4 rounded border-neutral-300 text-primary focus:ring-primary"
-            />
-            Enable a daily penalty window
-          </label>
-          <div className="grid gap-4 sm:grid-cols-2">
-            <label className="grid gap-1 text-sm text-neutral-600 dark:text-neutral-200">
-              <span className="text-xs font-semibold uppercase text-neutral-500">Start time</span>
-              <input
-                type="time"
-                value={penaltyStartTime}
-                onChange={(event) => setPenaltyStartTime(event.target.value)}
-                disabled={!penaltyDailyWindowEnabled}
-                className="rounded-lg border border-neutral-200 bg-white px-3 py-2 text-sm shadow-sm focus:border-primary focus:outline-none focus:ring-2 focus:ring-primary/40 disabled:cursor-not-allowed disabled:opacity-60 dark:border-midnight-700 dark:bg-midnight-900"
-              />
-            </label>
-            <label className="grid gap-1 text-sm text-neutral-600 dark:text-neutral-200">
-              <span className="text-xs font-semibold uppercase text-neutral-500">End time</span>
-              <input
-                type="time"
-                value={penaltyEndTime}
-                onChange={(event) => setPenaltyEndTime(event.target.value)}
-                disabled={!penaltyDailyWindowEnabled}
-                className="rounded-lg border border-neutral-200 bg-white px-3 py-2 text-sm shadow-sm focus:border-primary focus:outline-none focus:ring-2 focus:ring-primary/40 disabled:cursor-not-allowed disabled:opacity-60 dark:border-midnight-700 dark:bg-midnight-900"
-              />
-            </label>
-          </div>
-          <p className="text-xs text-neutral-500">
-            {penaltyDailyWindowEnabled
-              ? 'Time range applies to every day unless the day is configured as an all-day penalty below.'
-              : 'When disabled, no time of day automatically attracts penalty rates.'}
-          </p>
-        </fieldset>
-        <fieldset className="grid gap-3">
-          <legend className="text-xs font-semibold uppercase text-neutral-500">Penalty applies all day on</legend>
-          <div className="grid grid-cols-2 gap-2 sm:grid-cols-3">
-            {WEEKDAY_OPTIONS.map((option) => {
-              const checked = penaltyAllDayWeekdays.includes(option.value);
-              return (
-                <label key={option.value} className="flex items-center gap-2 text-sm text-neutral-600 dark:text-neutral-200">
+        </aside>
+        <div className="flex-1">
+          {activeTab === 'general' ? (
+            <form className="flex flex-col gap-5" onSubmit={handleSubmit}>
+              <div className="grid gap-5">
+                <div className="grid gap-2">
+                  <label className="text-xs font-semibold uppercase text-neutral-500">Base rate (per hour)</label>
+                  <input
+                    type="number"
+                    min="0"
+                    step="0.01"
+                    value={baseRate}
+                    onChange={(event) => setBaseRate(Number(event.target.value))}
+                    className="rounded-lg border border-neutral-200 bg-white px-3 py-2 text-sm shadow-sm focus:border-primary focus:outline-none focus:ring-2 focus:ring-primary/40 dark:border-midnight-700 dark:bg-midnight-900"
+                  />
+                </div>
+                <div className="grid gap-2">
+                  <label className="text-xs font-semibold uppercase text-neutral-500">Penalty rate (per hour)</label>
+                  <input
+                    type="number"
+                    min="0"
+                    step="0.01"
+                    value={penaltyRate}
+                    onChange={(event) => setPenaltyRate(Number(event.target.value))}
+                    className="rounded-lg border border-neutral-200 bg-white px-3 py-2 text-sm shadow-sm focus:border-primary focus:outline-none focus:ring-2 focus:ring-primary/40 dark:border-midnight-700 dark:bg-midnight-900"
+                  />
+                </div>
+                <div className="grid gap-2">
+                  <label className="text-xs font-semibold uppercase text-neutral-500" htmlFor="week-starts-on">
+                    Pay week starts on
+                  </label>
+                  <p id="week-starts-on-help" className="text-xs text-neutral-500 dark:text-neutral-300">
+                    Only affects summary and payslip alignment.
+                  </p>
+                  <select
+                    id="week-starts-on"
+                    aria-describedby="week-starts-on-help"
+                    value={weekStartsOn}
+                    onChange={(event) => setWeekStartsOn(Number(event.target.value) as WeekStart)}
+                    className="rounded-lg border border-neutral-200 bg-white px-3 py-2 text-sm shadow-sm focus:border-primary focus:outline-none focus:ring-2 focus:ring-primary/40 dark:border-midnight-700 dark:bg-midnight-900"
+                  >
+                    {WEEK_START_OPTIONS.map((option) => (
+                      <option key={option.value} value={option.value}>
+                        {option.label}
+                      </option>
+                    ))}
+                  </select>
+                </div>
+                <div className="grid gap-2">
+                  <label className="text-xs font-semibold uppercase text-neutral-500">Currency</label>
+                  <input
+                    type="text"
+                    value={currency}
+                    onChange={(event) => setCurrency(event.target.value.toUpperCase())}
+                    className="rounded-lg border border-neutral-200 bg-white px-3 py-2 text-sm uppercase shadow-sm focus:border-primary focus:outline-none focus:ring-2 focus:ring-primary/40 dark:border-midnight-700 dark:bg-midnight-900"
+                    maxLength={3}
+                  />
+                </div>
+              </div>
+              {renderFormActions()}
+            </form>
+          ) : null}
+
+          {activeTab === 'notifications' ? (
+            <form className="flex flex-col gap-5" onSubmit={handleSubmit}>
+              <fieldset className="grid gap-3">
+                <legend className="text-xs font-semibold uppercase text-neutral-500">Shift reminders</legend>
+                <p className="text-xs text-neutral-500 dark:text-neutral-300">
+                  Configure how far in advance the app reminds you about upcoming shifts.
+                </p>
+                <div className="grid gap-3 sm:grid-cols-3">
+                  <label className="grid gap-1 text-sm text-neutral-600 dark:text-neutral-200">
+                    <span className="text-xs font-semibold uppercase text-neutral-500">Long-range (minutes)</span>
+                    <input
+                      type="number"
+                      min={0}
+                      max={7 * 24 * 60}
+                      value={notificationLongLead}
+                      onChange={(event) => setNotificationLongLead(Number(event.target.value))}
+                      className="rounded-lg border border-neutral-200 bg-white px-3 py-2 text-sm shadow-sm focus:border-primary focus:outline-none focus:ring-2 focus:ring-primary/40 dark:border-midnight-700 dark:bg-midnight-900"
+                    />
+                    <span className="text-xs text-neutral-500 dark:text-neutral-300">
+                      Send a one-off reminder this many minutes before the shift.
+                    </span>
+                  </label>
+                  <label className="grid gap-1 text-sm text-neutral-600 dark:text-neutral-200">
+                    <span className="text-xs font-semibold uppercase text-neutral-500">Short-range (minutes)</span>
+                    <input
+                      type="number"
+                      min={0}
+                      max={24 * 60}
+                      value={notificationShortLead}
+                      onChange={(event) => setNotificationShortLead(Number(event.target.value))}
+                      className="rounded-lg border border-neutral-200 bg-white px-3 py-2 text-sm shadow-sm focus:border-primary focus:outline-none focus:ring-2 focus:ring-primary/40 dark:border-midnight-700 dark:bg-midnight-900"
+                    />
+                    <span className="text-xs text-neutral-500 dark:text-neutral-300">
+                      Start persistent reminders once the shift is this close.
+                    </span>
+                  </label>
+                  <label className="grid gap-1 text-sm text-neutral-600 dark:text-neutral-200">
+                    <span className="text-xs font-semibold uppercase text-neutral-500">Repeat every (minutes)</span>
+                    <input
+                      type="number"
+                      min={5}
+                      max={24 * 60}
+                      value={notificationRepeat}
+                      onChange={(event) => setNotificationRepeat(Number(event.target.value))}
+                      className="rounded-lg border border-neutral-200 bg-white px-3 py-2 text-sm shadow-sm focus:border-primary focus:outline-none focus:ring-2 focus:ring-primary/40 dark:border-midnight-700 dark:bg-midnight-900"
+                    />
+                    <span className="text-xs text-neutral-500 dark:text-neutral-300">
+                      How often short-range reminders repeat until the shift starts.
+                    </span>
+                  </label>
+                </div>
+              </fieldset>
+              {renderFormActions()}
+            </form>
+          ) : null}
+
+          {activeTab === 'penalties' ? (
+            <form className="flex flex-col gap-5" onSubmit={handleSubmit}>
+              <fieldset className="grid gap-3">
+                <legend className="text-xs font-semibold uppercase text-neutral-500">Penalty hours (daily window)</legend>
+                <label className="flex items-center gap-2 text-sm text-neutral-600 dark:text-neutral-200">
                   <input
                     type="checkbox"
-                    checked={checked}
-                    onChange={(event) => {
-                      setPenaltyAllDayWeekdays((current) => {
-                        if (event.target.checked) {
-                          return Array.from(new Set<Weekday>([...current, option.value] as Weekday[]));
-                        }
-                        return current.filter((day) => day !== option.value);
-                      });
-                    }}
+                    checked={penaltyDailyWindowEnabled}
+                    onChange={(event) => setPenaltyDailyWindowEnabled(event.target.checked)}
                     className="h-4 w-4 rounded border-neutral-300 text-primary focus:ring-primary"
                   />
-                  {option.label}
+                  Enable a daily penalty window
                 </label>
-              );
-            })}
-          </div>
-        </fieldset>
-        <fieldset className="grid gap-3">
-          <legend className="text-xs font-semibold uppercase text-neutral-500">Public holidays</legend>
-          <label className="flex items-center gap-2 text-sm text-neutral-600 dark:text-neutral-200">
-            <input
-              type="checkbox"
-              checked={includePublicHolidays}
-              onChange={(event) => setIncludePublicHolidays(event.target.checked)}
-              className="h-4 w-4 rounded border-neutral-300 text-primary focus:ring-primary"
-            />
-            Include public holidays as all-day penalty shifts
-          </label>
-          <div className="grid gap-2 sm:grid-cols-[minmax(0,_200px)]">
-            <label className="grid gap-1 text-sm text-neutral-600 dark:text-neutral-200">
-              <span className="text-xs font-semibold uppercase text-neutral-500">Holiday region</span>
-              <select
-                value={publicHolidayCountry}
-                onChange={(event) => setPublicHolidayCountry(event.target.value.toUpperCase())}
-                className="rounded-lg border border-neutral-200 bg-white px-3 py-2 text-sm shadow-sm focus:border-primary focus:outline-none focus:ring-2 focus:ring-primary/40 dark:border-midnight-700 dark:bg-midnight-900"
-                disabled={!includePublicHolidays}
-              >
-                {publicHolidayRegionOptions.map((option) => (
-                  <option key={option.code} value={option.code}>
-                    {option.label}
-                  </option>
-                ))}
-              </select>
-            </label>
-            {isLoadingRegions ? (
-              <p className="text-xs text-neutral-500">Loading regions…</p>
-            ) : null}
-            {holidayRegions.length > 0 ? (
-              <label className="grid gap-1 text-sm text-neutral-600 dark:text-neutral-200">
-                <span className="text-xs font-semibold uppercase text-neutral-500">State or region</span>
-                <select
-                  value={publicHolidaySubdivision}
-                  onChange={(event) => setPublicHolidaySubdivision(event.target.value)}
-                  className="rounded-lg border border-neutral-200 bg-white px-3 py-2 text-sm shadow-sm focus:border-primary focus:outline-none focus:ring-2 focus:ring-primary/40 dark:border-midnight-700 dark:bg-midnight-900"
-                  disabled={!includePublicHolidays || isLoadingRegions}
-                >
-                  <option value="">Whole country</option>
-                  {holidayRegions.map((region) => (
-                    <option key={region.code} value={region.code}>
-                      {region.name}
-                    </option>
-                  ))}
-                </select>
-              </label>
-            ) : null}
-            {regionsError ? <p className="text-xs text-red-500">{regionsError}</p> : null}
-          </div>
-          <p className="text-xs text-neutral-500">
-            Holiday dates are sourced from{' '}
-            <a
-              href="https://date.nager.at/"
-              target="_blank"
-              rel="noreferrer"
-              className="text-primary underline"
-            >
-              Nager.Date
-            </a>{' '}
-            and cached for offline use.
-          </p>
-          {includePublicHolidays && cachedHolidayCount > 0 ? (
-            <p className="text-xs text-emerald-600">
-              Cached {cachedHolidayCount} holidays for {publicHolidayCountry}
-              {selectedSubdivisionName ? ` (${selectedSubdivisionName})` : ''}.
-            </p>
+                <div className="grid gap-4 sm:grid-cols-2">
+                  <label className="grid gap-1 text-sm text-neutral-600 dark:text-neutral-200">
+                    <span className="text-xs font-semibold uppercase text-neutral-500">Start time</span>
+                    <input
+                      type="time"
+                      value={penaltyStartTime}
+                      onChange={(event) => setPenaltyStartTime(event.target.value)}
+                      disabled={!penaltyDailyWindowEnabled}
+                      className="rounded-lg border border-neutral-200 bg-white px-3 py-2 text-sm shadow-sm focus:border-primary focus:outline-none focus:ring-2 focus:ring-primary/40 disabled:cursor-not-allowed disabled:opacity-60 dark:border-midnight-700 dark:bg-midnight-900"
+                    />
+                  </label>
+                  <label className="grid gap-1 text-sm text-neutral-600 dark:text-neutral-200">
+                    <span className="text-xs font-semibold uppercase text-neutral-500">End time</span>
+                    <input
+                      type="time"
+                      value={penaltyEndTime}
+                      onChange={(event) => setPenaltyEndTime(event.target.value)}
+                      disabled={!penaltyDailyWindowEnabled}
+                      className="rounded-lg border border-neutral-200 bg-white px-3 py-2 text-sm shadow-sm focus:border-primary focus:outline-none focus:ring-2 focus:ring-primary/40 disabled:cursor-not-allowed disabled:opacity-60 dark:border-midnight-700 dark:bg-midnight-900"
+                    />
+                  </label>
+                </div>
+                <p className="text-xs text-neutral-500">
+                  {penaltyDailyWindowEnabled
+                    ? 'Time range applies to every day unless the day is configured as an all-day penalty below.'
+                    : 'When disabled, no time of day automatically attracts penalty rates.'}
+                </p>
+              </fieldset>
+
+              <fieldset className="grid gap-3">
+                <legend className="text-xs font-semibold uppercase text-neutral-500">Penalty applies all day on</legend>
+                <div className="grid grid-cols-2 gap-2 sm:grid-cols-3">
+                  {WEEKDAY_OPTIONS.map((option) => {
+                    const checked = penaltyAllDayWeekdays.includes(option.value);
+                    return (
+                      <label key={option.value} className="flex items-center gap-2 text-sm text-neutral-600 dark:text-neutral-200">
+                        <input
+                          type="checkbox"
+                          checked={checked}
+                          onChange={(event) => {
+                            setPenaltyAllDayWeekdays((current) => {
+                              if (event.target.checked) {
+                                return Array.from(new Set<Weekday>([...current, option.value] as Weekday[]));
+                              }
+                              return current.filter((day) => day !== option.value);
+                            });
+                          }}
+                          className="h-4 w-4 rounded border-neutral-300 text-primary focus:ring-primary"
+                        />
+                        {option.label}
+                      </label>
+                    );
+                  })}
+                </div>
+              </fieldset>
+
+              <fieldset className="grid gap-3">
+                <legend className="text-xs font-semibold uppercase text-neutral-500">Public holidays</legend>
+                <label className="flex items-center gap-2 text-sm text-neutral-600 dark:text-neutral-200">
+                  <input
+                    type="checkbox"
+                    checked={includePublicHolidays}
+                    onChange={(event) => setIncludePublicHolidays(event.target.checked)}
+                    className="h-4 w-4 rounded border-neutral-300 text-primary focus:ring-primary"
+                  />
+                  Include public holidays as all-day penalty shifts
+                </label>
+                <div className="grid gap-2 sm:grid-cols-[minmax(0,_200px)]">
+                  <label className="grid gap-1 text-sm text-neutral-600 dark:text-neutral-200">
+                    <span className="text-xs font-semibold uppercase text-neutral-500">Holiday region</span>
+                    <select
+                      value={publicHolidayCountry}
+                      onChange={(event) => setPublicHolidayCountry(event.target.value.toUpperCase())}
+                      className="rounded-lg border border-neutral-200 bg-white px-3 py-2 text-sm shadow-sm focus:border-primary focus:outline-none focus:ring-2 focus:ring-primary/40 dark:border-midnight-700 dark:bg-midnight-900"
+                      disabled={!includePublicHolidays}
+                    >
+                      {publicHolidayRegionOptions.map((option) => (
+                        <option key={option.code} value={option.code}>
+                          {option.label}
+                        </option>
+                      ))}
+                    </select>
+                  </label>
+                  {isLoadingRegions ? <p className="text-xs text-neutral-500">Loading regions…</p> : null}
+                  {holidayRegions.length > 0 ? (
+                    <label className="grid gap-1 text-sm text-neutral-600 dark:text-neutral-200">
+                      <span className="text-xs font-semibold uppercase text-neutral-500">State or region</span>
+                      <select
+                        value={publicHolidaySubdivision}
+                        onChange={(event) => setPublicHolidaySubdivision(event.target.value)}
+                        className="rounded-lg border border-neutral-200 bg-white px-3 py-2 text-sm shadow-sm focus:border-primary focus:outline-none focus:ring-2 focus:ring-primary/40 dark:border-midnight-700 dark:bg-midnight-900"
+                        disabled={!includePublicHolidays || isLoadingRegions}
+                      >
+                        <option value="">Whole country</option>
+                        {holidayRegions.map((region) => (
+                          <option key={region.code} value={region.code}>
+                            {region.name}
+                          </option>
+                        ))}
+                      </select>
+                    </label>
+                  ) : null}
+                  {regionsError ? <p className="text-xs text-red-500">{regionsError}</p> : null}
+                </div>
+                <p className="text-xs text-neutral-500">
+                  Holiday dates are sourced from{' '}
+                  <a href="https://date.nager.at/" target="_blank" rel="noreferrer" className="text-primary underline">
+                    Nager.Date
+                  </a>{' '}
+                  and cached for offline use.
+                </p>
+                {includePublicHolidays && cachedHolidayCount > 0 ? (
+                  <p className="text-xs text-emerald-600">
+                    Cached {cachedHolidayCount} holidays for {publicHolidayCountry}
+                    {selectedSubdivisionName ? ` (${selectedSubdivisionName})` : ''}.
+                  </p>
+                ) : null}
+              </fieldset>
+              {renderFormActions()}
+            </form>
           ) : null}
-        </fieldset>
-        <button
-          type="submit"
-          className="w-full rounded-full bg-primary px-5 py-2 text-sm font-semibold text-primary-foreground shadow transition hover:bg-midnight-900 disabled:cursor-not-allowed disabled:opacity-60 sm:w-auto"
-          disabled={isSaving}
-        >
-          {isSaving ? 'Saving…' : 'Save settings'}
-        </button>
-        {formError && <p className="text-xs text-red-500">{formError}</p>}
-        {status && <p className="text-xs text-emerald-500">{status}</p>}
-      </form>
-      <ImportExportModal isOpen={isImportExportOpen} onClose={() => setIsImportExportOpen(false)} />
-      <BackupRestoreModal isOpen={isBackupModalOpen} onClose={() => setIsBackupModalOpen(false)} />
+
+          {activeTab === 'appearance' ? (
+            <form className="flex flex-col gap-5" onSubmit={handleSubmit}>
+              <fieldset className="grid gap-3">
+                <legend className="text-xs font-semibold uppercase text-neutral-500">Theme</legend>
+                <div className="grid gap-2">
+                  {THEME_OPTIONS.map((option) => (
+                    <label
+                      key={option.value}
+                      className="flex items-start gap-2 rounded-lg border border-neutral-200 bg-white px-3 py-2 text-sm shadow-sm transition hover:border-primary/60 dark:border-midnight-700 dark:bg-midnight-900"
+                    >
+                      <input
+                        type="radio"
+                        name="theme"
+                        value={option.value}
+                        checked={theme === option.value}
+                        onChange={() => setTheme(option.value)}
+                        className="mt-1 h-4 w-4 border-neutral-300 text-primary focus:ring-primary"
+                      />
+                      <span className="flex flex-col">
+                        <span className="font-medium text-neutral-700 dark:text-neutral-100">{option.label}</span>
+                        <span className="text-xs text-neutral-500 dark:text-neutral-300">{option.description}</span>
+                      </span>
+                    </label>
+                  ))}
+                </div>
+              </fieldset>
+
+              <div className="grid gap-2">
+                <span className="text-xs font-semibold uppercase text-neutral-500">Time format</span>
+                <label className="flex items-center gap-2 text-sm text-neutral-600 dark:text-neutral-200">
+                  <input
+                    type="checkbox"
+                    checked={use24HourTime}
+                    onChange={(event) => setUse24HourTime(event.target.checked)}
+                    className="h-4 w-4 rounded border-neutral-300 text-primary focus:ring-primary"
+                  />
+                  Use 24-hour clock
+                </label>
+                <p className="text-xs text-neutral-500 dark:text-neutral-300">
+                  Applies to shift start and end times shown in the app.
+                </p>
+              </div>
+              {renderFormActions()}
+            </form>
+          ) : null}
+
+          {activeTab === 'data' ? (
+            <div className="space-y-8">
+              <section className="space-y-3">
+                <h3 className="text-base font-semibold text-neutral-900 dark:text-neutral-50">Import & export shifts</h3>
+                <p className="text-sm text-neutral-600 dark:text-neutral-300">
+                  Download a CSV of your shifts or import new ones using the Chrona template.
+                </p>
+                <ImportExportPanel />
+              </section>
+              <section className="space-y-3">
+                <h3 className="text-base font-semibold text-neutral-900 dark:text-neutral-50">Backup & restore</h3>
+                <p className="text-sm text-neutral-600 dark:text-neutral-300">
+                  Create a full Chrona backup or restore from a saved archive to migrate devices.
+                </p>
+                <BackupRestorePanel />
+              </section>
+            </div>
+          ) : null}
+        </div>
+      </div>
     </section>
   );
 }
